cart-service/requirements.txt
django
djangorestframework
requests
psycopg2-binary
pika
PyJWT
